apps/api-edge/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
# sqlcipher3==0.5.2 (Requires system headers, skipping for initial setup)
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
watchdog==4.0.0
bcrypt==4.1.2
